+++ packages/frontend/README.md
@@ -8,6 +8,8 @@
 
 ※ nvm等の他ツールがお好みであればnvm等でも可です
 
+↑ 最終的にDocker化したら何もかも不要になりそう
+
 # installs
 
 ## volta
